server/requirements.txt
eventlet==0.21.0
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Migrate==2.1.1
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
pytest==3.2.3
WTForms==2.1